requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
python-dotenv
djangorestframework-simplejwt
drf-spectacular
django-rest-swagger